Move common options things into firrtl.options
Topological refactor to put all options parsing related things (e.g.,
ExecutionOptionsManager) into the firrtl.options package.
