+++ src/test/scala/firrtlTests/InternalErrorSpec.scala
@@ -5,6 +5,7 @@ package firrtlTests
 import java.io.File
 
 import firrtl._
+import firrtl.options.ExecutionOptionsManager
 import firrtl.Utils.getThrowable
 import firrtl.util.BackendCompilationUtilities
 import org.scalatest.{FreeSpec, Matchers}
